Refactor main.py
- Extract dataset loading functions into the dataset module
- Move global variables into the config_wordalgebra module
